+++ Src/SLib/ssvg.cpp
@@ -382,7 +382,7 @@ int SSvg::ProcessStyleItem(int token, const SString & rVal, StyleBlock & rBlk)
 					temp_buf.ShiftLeftChr('(');
 					temp_buf.ShiftLeftChr('\"');
 					temp_buf.ShiftLeftChr('#');
-					uint par_pos = 0;
+					size_t par_pos = 0;
 					if(temp_buf.StrChr(')', &par_pos))
 						temp_buf.Trim(par_pos);
 					temp_buf.TrimRightChr('\"');
